backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.12
PyJWT==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
